requirements.txt
tensorflow
matplotlib
hyperopt
numpy
numpy-quaternion
